+++ magda-web-client/src/Components/Dataset/Add/DatasetAddMetadataPage.tsx
@@ -24,6 +24,7 @@ import DatasetAddPeoplePage from "./Pages/People/DatasetAddPeoplePage";
 import DatasetAddEndPreviewPage from "./Pages/DatasetAddEndPreviewPage";
 import DatasetAddEndPage from "./Pages/DatasetAddEndPage";
 import DatasetAddFilesPage from "./Pages/AddFiles";
+import ReviewPage from "./Pages/ReviewPage/index";
 import DatasetAddAccessAndUsePage from "./Pages/DatasetAddAccessAndUsePage";
 import withAddDatasetState from "./withAddDatasetState";
 import { config } from "config";
@@ -38,6 +39,7 @@ import ErrorMessageBox from "./ErrorMessageBox";
 import helpIcon from "assets/help.svg";
 import { User } from "reducers/userManagementReducer";
 import * as ValidationManager from "../Add/ValidationManager";
+import URI from "urijs";
 
 type Props = {
     initialState: State;
@@ -50,6 +52,7 @@ type Props = {
     datasetId: string;
     history: any;
     user: User;
+    isBackToReview: boolean;
 };
 
 class NewDataset extends React.Component<Props, State> {
@@ -100,15 +103,16 @@ class NewDataset extends React.Component<Props, State> {
                 isEditView={true}
             />
         ),
+        this.renderSubmitPage.bind(this),
+        () => <ReviewPage stateData={this.state} />,
         config.featureFlags.previewAddDataset
             ? () => <DatasetAddEndPreviewPage />
-            : this.renderSubmitPage.bind(this),
-        () => (
-            <DatasetAddEndPage
-                datasetId={this.props.datasetId}
-                history={this.props.history}
-            />
-        )
+            : () => (
+                  <DatasetAddEndPage
+                      datasetId={this.props.datasetId}
+                      history={this.props.history}
+                  />
+              )
     ];
 
     edit = <K extends keyof State>(aspectField: K) => (field: string) => (
@@ -132,32 +136,59 @@ class NewDataset extends React.Component<Props, State> {
 
         step = Math.max(Math.min(step, this.steps.length - 1), 0);
 
-        const nextIsPublish = step === 4;
+        const hideExitButton = config.featureFlags.previewAddDataset
+            ? step >= 4
+            : step >= 5;
 
         const nextButtonCaption = () => {
-            if (nextIsPublish) {
-                if (config.featureFlags.previewAddDataset) {
-                    return "Send Us Your Thoughts";
-                } else if (this.state.isPublishing) {
+            if (step === 5) {
+                // --- review page
+                if (this.state.isPublishing) {
                     return "Publishing as draft...";
                 } else {
                     return "Publish draft dataset";
                 }
+            } else if (step === 6) {
+                // --- all done or preview mode feedback page
+                // --- All done page has no button to show
+                return "Send Us Your Thoughts";
             } else {
-                return ProgressMeterStepsConfig[step + 1].title;
+                return "Next: " + ProgressMeterStepsConfig[step + 1].title;
             }
         };
 
         const nextButtonOnClick = () => {
-            if (nextIsPublish) {
+            if (step === 5) {
+                // --- review page
                 if (config.featureFlags.previewAddDataset) {
-                    window.location.href =
-                        "mailto:magda@csiro.au?subject=Add Dataset Feedback";
+                    this.gotoStep(step + 1);
                 } else {
                     this.performPublishDataset();
                 }
+            } else if (step === 6) {
+                // --- all done or preview mode feedback page
+                if (config.featureFlags.previewAddDataset) {
+                    // --- preview mode feedback page
+                    window.location.href =
+                        "mailto:magda@csiro.au?subject=Add Dataset Feedback";
+                }
             } else {
                 this.gotoStep(step + 1);
+            }
+        };
+
+        const shouldRenderButtonArea = () => {
+            if (
+                distributions.filter(
+                    item => item._state !== DistributionState.Ready
+                ).length
+            ) {
+                return false;
+            }
+            if (step === 6 && !config.featureFlags.previewAddDataset) {
+                return false;
+            } else {
+                return true;
             }
         };
 
@@ -180,29 +211,42 @@ class NewDataset extends React.Component<Props, State> {
                 <br />
                 <ErrorMessageBox />
                 <br />
-                {distributions.filter(
-                    item => item._state !== DistributionState.Ready
-                ).length ? null : (
+                {!shouldRenderButtonArea() ? null : (
                     <>
-                        {step >= 5 ? null : (
-                            <div className="row next-save-button-row">
-                                <div className="col-sm-12">
+                        <div className="row next-save-button-row">
+                            <div className="col-sm-12">
+                                {this.props.isBackToReview ? (
                                     <button
-                                        className="au-btn next-button"
-                                        onClick={nextButtonOnClick}
+                                        className="au-btn back-to-review-button"
+                                        onClick={() =>
+                                            this.gotoStep(this.steps.length - 2)
+                                        }
                                     >
-                                        Next: {nextButtonCaption()}
+                                        Return to Review
                                     </button>
-                                    {/* Only show the Save and Exit button on pages prior to step 4*/}
+                                ) : null}
+
+                                <button
+                                    className={`au-btn ${
+                                        this.props.isBackToReview
+                                            ? "au-btn--secondary save-button"
+                                            : "next-button"
+                                    }`}
+                                    onClick={nextButtonOnClick}
+                                    disabled={this.state.isPublishing}
+                                >
+                                    {nextButtonCaption()}
+                                </button>
+                                {hideExitButton ? null : (
                                     <button
                                         className="au-btn au-btn--secondary save-button"
                                         onClick={this.saveAndExit.bind(this)}
                                     >
                                         Save and exit
                                     </button>
-                                </div>
+                                )}
                             </div>
-                        )}
+                        </div>
                     </>
                 )}
             </div>
@@ -284,7 +328,7 @@ class NewDataset extends React.Component<Props, State> {
             );
 
             this.props.history.push(
-                `/dataset/add/metadata/${this.props.datasetId}/5`
+                `/dataset/add/metadata/${this.props.datasetId}/6`
             );
         } catch (e) {
             this.setState({
@@ -296,15 +340,20 @@ class NewDataset extends React.Component<Props, State> {
 }
 
 function mapStateToProps(state, props) {
-    const datasetId = props.match.params.datasetId;
+    const uri = new URI(location.href);
+    const datasetId = props?.match?.params?.datasetId;
     let step = parseInt(props?.match?.params?.step);
+    const isBackToReview =
+        typeof uri.search(true)?.isBackToReview !== "undefined" ? true : false;
+
     if (isNaN(step)) {
         step = 0;
     }
 
     return {
         datasetId,
-        step
+        step,
+        isBackToReview
     };
 }