+++ CHANGES.md
@@ -49,6 +49,7 @@ UI:
 -   Added the support of processing ESRI REST API URL
 -   Use IBM Plex Sans font
 -   Fixed: error notification no longer has a white background
+-   Added `add dataset review page`
 
 Storage:
 
@@ -87,6 +88,7 @@ Authorization:
 Search
 
 -   Began work on a NodeJS-based search API to replace the scala one
+-   Allow fetch region record by region id
 
 Others: